Address PR feedback: consolidate date/time utils
